server/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-appconf==1.0.3
django-axes==5.3.0
django-filter==2.2.0
django-ipware==2.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.1
wrapt==1.11.2
